Create initial AbstractAppliance
The AbstractAppliance is an "abstract" class which all TV Kitchen
appliances are expected to extend.  It provides:

- payload management
- payload validation flows
- event emission

This list will likely expand as we better understand the space.

It is important to note that because JavaScript does not officially
support the concept of abstract classes, abstraction is simulated by
defining methods that are expected to be overloaded.

Issue #1