Create AbstractInstantitationError
The AbstractInstantiationError will be used as we define our abstract
class.

AbstractInstantiationErrors are thrown when an abstract class is
instantiated (e.g. 'new AbstractClass()').  This is intended as
a workaround to the fact that JavaScript doesn't support the concept of
abstract classes yet.

This commit also makes an index.js for loading these errors with nicer
syntactic sugar.

Issue #1