Move to monorepo structure
We are moving to a monorepo structure for appliances, rather than having
individual appliances in their own separate repository. This will allow
others to more easily view a list of official TV Kitchen appliances,
and should make it easier to manage changes that span multiple
appliances.

For now we are choosing to avoid lerna, though that may change as we
develop better practices for publishing.

Issue #12 Convert this to a monorepo