Add isValidPayload to AbstractAppliance
After realizing that every single appliance is going to want to verify
that a Payload is a valid input type, we decided to add that logic to
the AbstractAppliance.

The decision to use an assertion here (thus triggering an error) is
intentional, since submitting a payload of an invalid type for
validation is likely a signal of a more serious problem.

Issue #41