Always use named exports
We never want to use default exports. There are a few reasons for this.
One is that Babel doesn't actually map them to default, which means that
require() doesn't work as expected.

The more important reason is that by using named exports we create
a consistant expectation for how to use appliances and other TVK
components, and also allow appliances to export supporting classes as
necessary.

Issue #55