Support timestamp in video ingestor appliances
We were ignoring the concept of timestamps because they're tricky, but
this adds it in the AbstractVideoIngestionAppliance, which means it adds
it for all ingestion appliances that exist so far.

Implementations of TV Kitchen can pass an `origin` setting which
indicates what timestamp position 0 is associated with.  This allows for
use cases such as processing a pre-recorded video file that captures
a moment in time.

Issue #76