Allow null invoke returns
The `ingestPayload` method calls `invoke`, which is documented as
returning a PayloadArray.  This is fine, but we need to allow it to
return null for a few reasons.  First, it's a bit odd to fail if
a function returns null, but second -- null is a much more convenient
way to indicate that there are no leftover payloads to process than
creating and returning an empty PayloadArray.

Issue #74