Add segment support to SRTs
Now that we have the concept of video segments, the SRT generator should
honor that concept, resetting its counter and origin time when a new
segment is announced.

As implemented, any topology that generates segments will trigger resets
in the SRT appliance / it is not an optional feature.  We can add that
option later in various ways -- for instance at the counterop level we
will likely want to allow a given appliance to filter a given station.

It also might make sense to add an appliance-level configuration
regarding segmentation in SRTs.

Issue #92