Refactor segment generator
The VideoSegmentGenerator appliance was using the old `timestamp` field
and needed to be updated to use `origin`.  This also helped highlight
some of the difficult-to-parse variable and method names.

Instead of using `origin` in the segment appliance, it is `startingAt`
which has more intuitive meaning.  Some of the internal methods have
been reworded as well with the goal of legibility.

Issue #103