Ignore stderr in CCExtractor process
CCExtractor doesn't generate too much noise in the stderr pipe, but it
does generate *some*.  This means that after a few hours the pipeline
stops accepting data.

This happened before as well in our ffmpeg process, so we just need to
pass the stdio map to the `spawn` command.  For now we'll just ignore
the output, though some day we might decide to consume and log it via
the appliance `logger`.

Issue #112