Remove hooks from base package
These hooks just meant yarn would regularly build everything, which
was completely pointless since the tests are run on pre-transpiled code
as is the linting.

Prepublish hooks never got invoked because our publication flow involves
going into the package we're publishing, which is out of the scope of
the hook.

Issue #122