Update basetime to be a fraction
MPEG-TS times are in terms of 1/90000th of a second, and it seems that
value is written in terms of the full fraction, not just the
denominator.

We were storing the "90000" and dividing by it.  This change should
bring things in line with what I understand to be the more explicit way
of referencing basetime.