Only track video streams
The AbstractVideoIngestionAppliance keeps track of the position associated
with the stream; however, in MPEG-TS there is more than one stream (e.g.
audio and video) which means there is actually more than one stream
position.

This was causing issues where position was jumping around back and forth
because all positions were treated as being from a single stream.  We
now only care about the video stream.

It may be that an MPEG-TS stream can contain multiple video streams,
which will break this code; that said, TV Kitchen is fairly built around
an assumption that a given video stream has a single video in it.

In the long run we may want to also track audio stream position, but
honestly at that stage we would probably want the ingestor to output
multiple demuxed and remuxed streams instead of simply forwarding the
stream.

Issue #132