Rename AbstractVideoIngestionAppliance
Our appliances are not named in terms of the action taken but rather the
name of the actor.

The logic is that you would call something a "Toaster" instead of a "Toasting"

Anyway, Ingestion Appliances broke this approach, so we're renaming them
to Receivers.  This is a fairly apt name, since they serve a similar
function to what a TV or Radio receiver does -- i.e. taking the external
source of content and converting it into a standard format.

Changing the name of the abstract appliance is the first step towards
changing the name of the implemented appliances.

Issue #136