Load CCExtractor lines all at once
The previous caption extractor logic attempted to parse out CCExtractor
captions as they were generated.  This had the benefit of getting the
data as fast as possible, but in reality this could create a lot of
downstream complexity because captions have a mechanism for correction.

For instance, if a stenographer of a live program makes a correction to
a word we would require downstream appliances to know how to remove the
previous data and potentially even issue corrections of their own to
their downstream.

Things get much more simple if we only extract finalized lines.  This
introduces a minor delay since no content is rendered until the line's
completion, but for now that is a trade-off we are comfortable with.

Issue #133