Support dynamic input / output types
Appliances define static `getInputType` and `getOutputType` methods.
This needs to be static because the Countertop needs to generate a toplogy
*before* instantiating any specific appliances.

By passing those methods the settings associated with a future
instantiation we make it possible for the methods to base the returned
types on the value of the settings.

Issue #151